requirements.txt
discord.py
matplotlib
python-dotenv
flask
